+++ framework/py/flwr/supercore/error/base.py
@@ -149,10 +149,7 @@ class ApiErrorCode(IntEnum):
     ACCOUNT_AUTHENTICATION_NOT_INITIALIZED = 37
     INVALID_CONNECTOR_REQUEST = 38
     CONNECTOR_NOT_FOUND = 39
-    CONNECTOR_OAUTH_SESSION_NOT_FOUND = 40
-    CONNECTOR_OAUTH_SESSION_INVALID = 41
-    CONNECTOR_OAUTH_PROVIDER_FAILURE = 42
-    CONNECTOR_PERSISTENCE_FAILURE = 43
+    CONNECTOR_FAILURE = 40
 
     # Common API errors (1001-2000)
     RUNTIME_VERSION_INCOMPATIBLE = 1001
