+++ framework/py/flwr/superlink/servicer/control/control_handlers.py
@@ -16,10 +16,14 @@
 
 # pylint: disable=too-many-lines
 
+import base64
 import hashlib
 import json
-from collections.abc import Sequence
+import secrets
+from collections.abc import Callable, Sequence
+from datetime import datetime, timedelta
 from logging import ERROR, INFO
+from typing import NoReturn, TypeVar
 
 import requests
 
@@ -51,18 +55,27 @@ from flwr.proto.control_pb2 import (  # pylint: disable=E0611
     AddNodeToFederationResponse,
     ArchiveFederationRequest,
     ArchiveFederationResponse,
+    BeginConnectorOAuthRequest,
+    BeginConnectorOAuthResponse,
+    CompleteConnectorOAuthRequest,
+    CompleteConnectorOAuthResponse,
     ConfigureSimulationFederationRequest,
     ConfigureSimulationFederationResponse,
+    Connector,
     CreateFederationRequest,
     CreateFederationResponse,
     CreateInvitationRequest,
     CreateInvitationResponse,
+    DisconnectConnectorRequest,
+    DisconnectConnectorResponse,
     GetAuthTokensRequest,
     GetAuthTokensResponse,
     GetLoginDetailsRequest,
     GetLoginDetailsResponse,
     GetRunSeriesRequest,
     GetRunSeriesResponse,
+    ListConnectorsRequest,
+    ListConnectorsResponse,
     ListFederationsRequest,
     ListFederationsResponse,
     ListInvitationsRequest,
@@ -113,6 +126,8 @@ from flwr.supercore.error import ApiErrorCode, FlowerError
 from flwr.supercore.fab import Fab
 from flwr.supercore.primitives.asymmetric import bytes_to_public_key, uses_nist_ec_curve
 from flwr.supercore.run import Run
+from flwr.supercore.task_process.connector import registry as connector_registry
+from flwr.supercore.task_process.connector.oauth import ConnectorOAuthProvider
 from flwr.supercore.typing import (
     AcceptInvitationContext,
     CreateFederationContext,
@@ -124,10 +139,302 @@ from flwr.supercore.utils import (
     parse_app_spec,
     request_download_link,
     resolve_account_ids,
+    strict_json_dumps,
 )
 from flwr.superlink.artifact_provider import ArtifactProvider
 from flwr.superlink.auth_plugin import ControlAuthnPlugin
 from flwr.superlink.federation.noop_federation_manager import NoOpFederationManager
+
+OAUTH_SESSION_TTL = timedelta(minutes=10)
+T = TypeVar("T")
+
+
+def list_connectors(
+    request: ListConnectorsRequest,
+    account: AccountInfo,
+    state: LinkState,
+) -> ListConnectorsResponse:
+    """List user-connectable OAuth providers and account connection status."""
+    log(INFO, "ControlServicer.ListConnectors")
+    _ = request
+
+    connectors: list[Connector] = []
+    for provider in sorted(
+        connector_registry._OAUTH_CONNECTOR_PROVIDERS,
+        key=lambda item: item.connector_ref,
+    ):
+        connector_ref = provider.connector_ref
+        try:
+            connected = (
+                state.get_connector(
+                    flwr_aid=account.flwr_aid, connector_ref=connector_ref
+                )
+                is not None
+            )
+        except Exception as err:  # Persistence boundary
+            raise FlowerError(
+                ApiErrorCode.CONNECTOR_FAILURE,
+                f"Failed to list connectors ({type(err).__name__}).",
+            ) from None
+        connectors.append(
+            Connector(
+                connector_ref=connector_ref,
+                display_name=provider.display_name,
+                description=provider.description,
+                connected=connected,
+            )
+        )
+    return ListConnectorsResponse(connectors=connectors)
+
+
+def disconnect_connector(
+    request: DisconnectConnectorRequest,
+    account: AccountInfo,
+    state: LinkState,
+) -> DisconnectConnectorResponse:
+    """Delete one account-scoped connector connection."""
+    log(INFO, "ControlServicer.DisconnectConnector")
+    connector_ref = request.connector_ref.strip().lower()
+    if not connector_ref:
+        _raise_invalid_connector_request("connector_ref is required")
+    _get_connector_oauth_provider(connector_ref)
+
+    deleted = _call_connector_state(
+        lambda: state.delete_connector(
+            flwr_aid=account.flwr_aid, connector_ref=connector_ref
+        ),
+        "disconnect connector",
+    )
+    if not deleted:
+        raise FlowerError(
+            ApiErrorCode.CONNECTOR_NOT_FOUND,
+            f"Connector '{connector_ref}' is not connected for this account.",
+        )
+    return DisconnectConnectorResponse()
+
+
+def begin_connector_oauth(
+    request: BeginConnectorOAuthRequest,
+    account: AccountInfo,
+    state: LinkState,
+) -> BeginConnectorOAuthResponse:
+    """Create a short-lived account-scoped OAuth session."""
+    log(INFO, "ControlServicer.BeginConnectorOAuth")
+    connector_ref = request.connector_ref.strip().lower()
+    if not connector_ref:
+        _raise_invalid_connector_request("connector_ref is required")
+    redirect_uri = request.redirect_uri.strip()
+    if not redirect_uri:
+        _raise_invalid_connector_request("redirect_uri is required")
+    provider = _get_connector_oauth_provider(connector_ref)
+    try:
+        redirect_uri = provider.resolve_redirect_uri(redirect_uri)
+    except ValueError:
+        _raise_invalid_connector_request(
+            "redirect_uri is not allowed for this connector"
+        )
+    except Exception as err:  # Provider boundary
+        _raise_connector_provider_failure(
+            connector_ref, "resolve redirect URI", type(err).__name__
+        )
+    if not redirect_uri:
+        _raise_connector_provider_failure(
+            connector_ref, "resolve redirect URI", "empty response"
+        )
+
+    oauth_session_id = secrets.token_urlsafe(32)
+    oauth_state = secrets.token_urlsafe(32)
+    pkce_verifier = secrets.token_urlsafe(64)
+    digest = hashlib.sha256(pkce_verifier.encode("ascii")).digest()
+    pkce_challenge = base64.urlsafe_b64encode(digest).rstrip(b"=").decode("ascii")
+    expires_at = now() + OAUTH_SESSION_TTL
+    authorization_url = _call_connector_provider(
+        lambda: provider.build_authorization_url(
+            redirect_uri=redirect_uri,
+            state=oauth_state,
+            pkce_challenge=pkce_challenge,
+        ),
+        connector_ref,
+        "build authorization URL",
+    )
+    if not authorization_url:
+        _raise_connector_provider_failure(
+            connector_ref, "build authorization URL", "empty response"
+        )
+
+    session = _call_connector_state(
+        lambda: state.create_connector_oauth_session(
+            oauth_session_id=oauth_session_id,
+            flwr_aid=account.flwr_aid,
+            connector_ref=connector_ref,
+            state=oauth_state,
+            redirect_uri=redirect_uri,
+            pkce_verifier=pkce_verifier,
+            expires_at=expires_at,
+        ),
+        "create OAuth session",
+    )
+    if session is None:
+        _raise_connector_failure("OAuth session could not be created")
+
+    return BeginConnectorOAuthResponse(
+        oauth_session_id=session.oauth_session_id,
+        authorization_url=authorization_url,
+        connector_ref=session.connector_ref,
+        expires_at=session.expires_at,
+    )
+
+
+def complete_connector_oauth(
+    request: CompleteConnectorOAuthRequest,
+    account: AccountInfo,
+    state: LinkState,
+) -> CompleteConnectorOAuthResponse:
+    """Exchange an OAuth code and persist one account-scoped connection."""
+    log(INFO, "ControlServicer.CompleteConnectorOAuth")
+    oauth_session_id = request.oauth_session_id.strip()
+    if not oauth_session_id:
+        _raise_invalid_connector_request("oauth_session_id is required")
+    authorization_code = request.code.strip()
+    if not authorization_code:
+        _raise_invalid_connector_request("code is required")
+    if not request.state:
+        _raise_invalid_connector_request("state is required")
+
+    session = _call_connector_state(
+        lambda: state.get_connector_oauth_session(
+            oauth_session_id=oauth_session_id, flwr_aid=account.flwr_aid
+        ),
+        "get OAuth session",
+    )
+    if session is None:
+        raise FlowerError(
+            ApiErrorCode.CONNECTOR_NOT_FOUND,
+            "Connector OAuth session was not found for this account.",
+        )
+
+    try:
+        expires_at = datetime.fromisoformat(session.expires_at)
+    except ValueError:
+        _raise_connector_failure("OAuth session expiry is invalid")
+    if expires_at.utcoffset() is None:
+        _raise_connector_failure("OAuth session expiry is timezone-naive")
+    if (
+        session.completed_at is not None
+        or expires_at <= now()
+        or not secrets.compare_digest(
+            request.state.encode("utf-8"), session.state.encode("utf-8")
+        )
+    ):
+        _raise_invalid_connector_request(
+            f"OAuth session '{session.oauth_session_id}' is invalid or no longer "
+            "pending"
+        )
+
+    connector_ref = session.connector_ref.strip().lower()
+    provider = _get_connector_oauth_provider(connector_ref)
+    claimed = _call_connector_state(
+        lambda: state.complete_connector_oauth_session(
+            oauth_session_id=session.oauth_session_id,
+            flwr_aid=account.flwr_aid,
+        ),
+        "complete OAuth session",
+    )
+    if not claimed:
+        _raise_invalid_connector_request(
+            f"OAuth session '{session.oauth_session_id}' is invalid or no longer "
+            "pending"
+        )
+
+    credentials, config = _call_connector_provider(
+        lambda: provider.exchange_code(
+            code=authorization_code,
+            redirect_uri=session.redirect_uri,
+            pkce_verifier=session.pkce_verifier,
+        ),
+        connector_ref,
+        "exchange authorization code",
+    )
+    try:
+        credentials_json = strict_json_dumps(credentials, compact=True)
+        config_json = strict_json_dumps(config, compact=True)
+    except (TypeError, ValueError) as err:
+        _raise_connector_provider_failure(
+            connector_ref,
+            "serialize exchanged credentials",
+            type(err).__name__,
+        )
+
+    stored = _call_connector_state(
+        lambda: state.upsert_connector(
+            flwr_aid=account.flwr_aid,
+            connector_ref=connector_ref,
+            credentials_json=credentials_json,
+            config_json=config_json,
+        ),
+        "store connector credentials",
+    )
+    if not stored:
+        _raise_connector_failure("Connector credentials could not be stored")
+    return CompleteConnectorOAuthResponse(connector_ref=connector_ref)
+
+
+def _get_connector_oauth_provider(connector_ref: str) -> ConnectorOAuthProvider:
+    """Return the registered OAuth provider for a connector reference."""
+    for provider in connector_registry._OAUTH_CONNECTOR_PROVIDERS:
+        if provider.connector_ref == connector_ref:
+            return provider
+    raise FlowerError(
+        ApiErrorCode.CONNECTOR_NOT_FOUND,
+        f"OAuth provider for connector '{connector_ref}' was not found.",
+    )
+
+
+def _call_connector_provider(
+    operation: Callable[[], T], connector_ref: str, action: str
+) -> T:
+    """Call provider code without exposing provider exception details."""
+    try:
+        return operation()
+    except Exception as err:  # Provider boundary
+        _raise_connector_provider_failure(connector_ref, action, type(err).__name__)
+
+
+def _call_connector_state(operation: Callable[[], T], action: str) -> T:
+    """Call connector state code and translate persistence failures."""
+    try:
+        return operation()
+    except Exception as err:  # Persistence boundary
+        raise FlowerError(
+            ApiErrorCode.CONNECTOR_FAILURE,
+            f"Failed to {action} ({type(err).__name__}).",
+        ) from None
+
+
+def _raise_invalid_connector_request(reason: str) -> NoReturn:
+    """Raise a sanitized invalid connector request error."""
+    raise FlowerError(
+        ApiErrorCode.INVALID_CONNECTOR_REQUEST,
+        f"Invalid connector request: {reason}.",
+    )
+
+
+def _raise_connector_provider_failure(
+    connector_ref: str, action: str, error_type: str
+) -> NoReturn:
+    """Raise a provider failure without including provider exception details."""
+    raise FlowerError(
+        ApiErrorCode.CONNECTOR_FAILURE,
+        f"Connector '{connector_ref}' failed to {action} ({error_type}).",
+    ) from None
+
+
+def _raise_connector_failure(reason: str) -> NoReturn:
+    """Raise a sanitized connector failure."""
+    raise FlowerError(
+        ApiErrorCode.CONNECTOR_FAILURE,
+        f"Connector failure: {reason}.",
+    )
 
 
 def start_run(  # pylint: disable=too-many-locals, too-many-statements
